Move HardwareVersion into separate module
